+++ src/components/DownloadButton/index.tsx
@@ -1,45 +1,20 @@
-import Image from 'next/image'
+import DownloadIcon from '../../../public/images/icons/download.svg'
 
 import styles from './styles.module.css'
 
 interface Props {
-  type: 'appstore' | 'googleplay'
+  text: string
 }
 
-export default function DownloadButton({ type }: Props) {
+export default function DownloadButton({ text }: Props) {
   return (
-    <a
-      href={
-        type === 'appstore'
-          ? 'https://apps.apple.com/br/app/que-barbada/id1598991618'
-          : 'https://play.google.com/store/apps/details?id=com.quebarbada.quebarbada'
-      }
-      target='_blank'
-      rel='noreferrer'
-    >
-      {type === 'appstore' ? (
-        <div className={styles.button}>
-          <Image
-            priority
-            layout='responsive'
-            width={992}
-            height={325}
-            src='/images/brands/appstore.png'
-            alt='appstore'
-          />
+    <a href='https://www.google.com/' target='_blank' rel='noreferrer'>
+      <div className={styles.container}>
+        <div className={styles.content}>
+          <DownloadIcon />
+          <p className={styles.text}>{text}</p>
         </div>
-      ) : (
-        <div className={styles.button}>
-          <Image
-            priority
-            layout='responsive'
-            width={992}
-            height={325}
-            src='/images/brands/googleplay.png'
-            alt='googleplay'
-          />
-        </div>
-      )}
+      </div>
     </a>
   )
 }
