+++ src/components/DownloadButton/styles.module.css
@@ -1,17 +1,39 @@
-.button {
-  position: relative;
-  height: auto;
-  min-height: 0;
-  width: 8.75rem;
+.container {
+  display: flex;
+
+  align-items: center;
+  justify-content: center;
+
+  width: 15rem;
+  height: 2.5rem;
+
+  background-color: var(--brand-color-purple);
+
+  border-radius: 10px;
+
+  box-shadow: 1px 2px 2px rgba(0, 0, 0, 0.25);
+
+  transition: 0.2s;
 }
 
-@media screen and (min-width: 768px) {
-  .button {
-    width: 13.5rem;
-  }
+.container:hover {
+  opacity: 0.8;
 }
-@media screen and (min-width: 1024px) {
-  .button {
-    width: 9rem;
-  }
+
+.content {
+  display: flex;
+  align-items: center;
+}
+
+.content svg {
+  margin-right: 0.5rem;
+}
+
+.text {
+  font-family: 'Roboto', sans-serif;
+  font-weight: 400;
+  font-size: 1.125rem;
+  line-height: 1.375rem;
+
+  color: var(--brand-color-white);
 }
